refactor: 경매 등록 유형 클래스 수정 (#CHZZ-35)
