+++ src/main/java/org/chzz/market/domain/auction/enums/AuctionRegisterType.java
@@ -1,15 +1,15 @@
-package org.chzz.market.domain.auction.dto;
+package org.chzz.market.domain.auction.enums;
 
 import lombok.Getter;
 
 @Getter
-public enum AuctionType {
+public enum AuctionRegisterType {
     PRE_REGISTER("사전 등록"),
     REGISTER("경매 등록");
 
     private final String description;
 
-    AuctionType(String description) {
+    AuctionRegisterType(String description) {
         this.description = description;
     }
